Time: 63 ms (86.04%), Space: 46.1 MB (75.22%) - LeetHub

+++ 1508-range-sum-of-sorted-subarray-sums/1508-range-sum-of-sorted-subarray-sums.java
@@ -17,6 +17,6 @@ class Solution {
             ans = (ans + newArr[left-1]);
             left++;
         }
-        return (int)((double)ans%(1e9+7));
+        return (int)(ans%(1e9+7));
     }
 }
